Update renewables output, cluster capacity
The previous version of the database listed the output as of
some renewables technologies incorrectly as ELCP -- these outputs
were updated to ELCP_Renewables.

Added a new table to the database called tech_new_cluster that
contains the typical capacities (in GW) of indidivual power plants
within a new technology cluster. This table is indexed by region and
technology type.