Add TOC extension to jupyter notebooks
Added the table of contents extension to jupyter notebooks. This is reflected
in changes to the README file that shows how to enable the extension at the
command line within the temoa-py3 environment. Updated the postBuild file
to reflect this extension, so that the TOC is enabled within Binder.
Also updated the notebooks to remove the manual TOC and numbering for each
section.