Add WASTE_BIO placeholder, update AGR to AGR_BIO
Added a new placeholder commodity called WASTE_BIO to reflect
waste biomass supply. Data for this commodity is to be updated, and
placeholder values and connections are based on the same structure
as the AGR commodity. The MaxActivity values for this commodity are
set to be a small value so that they don't impact the model solution
at this time.
Further, the AGR commodity is updated to AGR_BIO to better reflect
what it represents (agricultural residues).