Add H2 distribution costs, alternate fuels to transport (other) demand
Hydrogen distribution costs are updated based on the IEA 2019 Future of
Hydrogen report. Intra-region distribution is assumed to be 250 km in a 100 tpd
pipeline or 500 km in a 500 tpd pipeline, both of which correspond to ~$0.5/kg H2.

The 'other' demand category for diesel in transport (TMDOH_DSL) was updated to
include biodiesel and T_Liquids as inputs, other than B20 and diesel.