Add files via upload
This file contains a summary of the findings from Year 1 of the Open Energy Outlook initiative. The report focuses on a series of policy scenarios in our efforts to achieve deep reduction in carbon dioxide emissions by 2050. The report also contains results on modeling to generate alternatives in the net-zero carbon dioxide emissions policy scenario.